babel (babel v0.3.11)

This module acts as entry point for a number of Babel features and provides some of the riakc_pb_socket module functions adapted for babel datatypes.

### Working with Babel Datatypes

### Working with Reliable Workflows #### Workflow aware functions A workflow aware function is a function that schedules its execution when it is called inside a workflow context. Several functions in this module are workflow aware e.g. put/5, delete/3.
### Working with Babel Indices

Types

bucket/0

-type bucket() :: binary().

bucket_and_type/0

-type bucket_and_type() :: {bucket_type(), bucket()}.

bucket_type/0

-type bucket_type() :: binary().

datatype/0

-type datatype() :: babel_map:t() | babel_set:t() | babel_counter:t().

delete_opts/0

-type delete_opts() ::
    #{connection => pid(),
      connection_pool => atom(),
      r => quorum(),
      pr => quorum(),
      w => quorum(),
      dw => quorum(),
      pw => quorum(),
      n_val => non_neg_integer(),
      sloppy_quorum => boolean(),
      timeout => timeout(),
      '$delete_validated' => boolean()}.

get_opts/0

-type get_opts() ::
    #{connection => pid(),
      connection_pool => atom(),
      r => quorum(),
      pr => quorum(),
      cache => boolean(),
      cache_force_refresh => boolean(),
      if_modified => binary(),
      notfound_ok => boolean(),
      n_val => non_neg_integer(),
      basic_quorum => boolean(),
      sloppy_quorum => boolean(),
      head => boolean(),
      deletedvclock => boolean(),
      timeout => timeout(),
      '$get_validated' => boolean()}.

opts/0

-type opts() :: get_opts() | put_opts() | delete_opts().

put_opts/0

-type put_opts() ::
    #{connection => pid(),
      connection_pool => atom(),
      cache => boolean(),
      w => quorum(),
      dw => quorum(),
      pw => quorum(),
      if_not_modified => boolean(),
      if_none_match => boolean(),
      notfound_ok => boolean(),
      n_val => non_neg_integer(),
      sloppy_quorum => boolean(),
      return_body => boolean(),
      return_head => boolean(),
      timeout => timeout(),
      '$put_validated' => boolean()}.

quorum/0

-type quorum() :: non_neg_integer() | one | all | quorum | default.

riak_op/0

-type riak_op() :: riakc_datatype:update(term()).

type_spec/0

-type type_spec() :: babel_map:type_spec() | babel_set:type_spec() | babel_counter:type_spec().

Functions

create_index(Index, Collection)

-spec create_index(Index :: babel_index:t(), Collection :: babel_index_collection:t()) ->
                {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Calls create_index/3 passing the default options as third argument.

create_index(Index, Collection, Opts)

-spec create_index(Index :: babel_index:t(), Collection :: babel_index_collection:t(), Opts :: opts()) ->
                {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Schedules the creation of an index and its partitions according to Config using Reliable.

The updated collection is returned under the key result of the reliable:wf_result().
?> This function uses a workflow, see workflow/2 for an explanation of the possible return values.

delete(TypedBucket, Key, Opts)

-spec delete(TypedBucket :: bucket_and_type(), Key :: binary(), Opts :: delete_opts()) ->
          ok | {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

?> This function is workflow aware

drop_all_indices(Collection)

-spec drop_all_indices(Collection :: babel_index_collection:t()) ->
                    {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Calls drop_all_indices/3

drop_all_indices(Collection, Opts)

-spec drop_all_indices(Collection :: babel_index_collection:t(), Opts :: opts()) ->
                    {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Drops all indices in collection by calling drop_indices/3.

drop_index(Index, Collection0)

-spec drop_index(Index :: binary(), Collection0 :: babel_index_collection:t()) ->
              {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

drop_index(Index, Collection, Opts)

-spec drop_index(Index :: binary(), Collection :: babel_index_collection:t(), Opts :: opts()) ->
              {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Schedules the removal of the index with name IndexName from collection Collection and all its index partitions from Riak KV. In case the collection is itself being dropped by a parent workflow, the collection will not be updated in Riak.
?> This function uses a workflow, see workflow/2 for an explanation of the possible return values.

drop_indices(IdxNames, Collection)

-spec drop_indices(IdxNames :: [binary()], Collection :: babel_index_collection:t()) ->
                {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Calls drop_indices/3

drop_indices(IdxNames, Collection, Opts)

-spec drop_indices(IdxNames :: [binary()], Collection :: babel_index_collection:t(), Opts :: opts()) ->
                {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Schedules the removal from Riak KV of indices with names IdxNames from collection Collection and their respective partitions.
?> This function uses a workflow, see workflow/2 for an explanation of the possible return values.

execute(Poolname, Fun, Opts)

-spec execute(Poolname :: atom(),
        Fun :: fun((RiakConn :: pid()) -> Result :: any()),
        Opts :: riak_pool:opts()) ->
           {ok, Result :: any()} | {error, Reason :: any()} | no_return().

Executes a number of operations using the same Riak client connection provided by riak_pool app. Poolname must be an already started pool or the atom undefined. In case of the latter the map Options should have an entry for the connection key.

Options:
* timeout - time to get a connection from the pool * connection - the pid to use instead of getting a new one from the pool

get(TypedBucket, Key, Spec, Opts)

-spec get(TypedBucket :: bucket_and_type(), Key :: binary(), Spec :: type_spec(), Opts :: get_opts()) ->
       {ok, Datatype :: datatype()} | {error, Reason :: term()}.

Retrieves a Riak Datatype (counter, set or map) from bucket type and bucket TypedBucket and key Key. It uses type spec Spec to transform the Riak Datatype into a Babel Datatype and if successful returns {ok, Datatype} where Datatype is one of babel_counter, babel_set or babel_map. Returns {error, not_found} if the key is not on the server.

This function gets the riak client connection from the options Opts under the key connection which can have the connection pid or a function object returning a connection pid. This allows a lot of flexibility such as reusing a given connection over several calls and using your own connection pool and management.
In case the connection option does not provide a connection as explained above, this function tries to get a connection from the the default riak_pool connection pool if it was enabled through Babel's configuration options.

get_connection()

-spec get_connection() -> undefined | pid().

Returns a Riak connection managed by riak_pool from the process dictonary or undefined if there is none.

module(Term)

-spec module(Term :: any()) -> module() | undefined.

Returns the module associated with the type of term Term.

opts_to_riak_opts(Opts)

-spec opts_to_riak_opts(map()) -> list().

Converts the options map into Riak KV property list format. It fails with a badarg exception if Opts is not the result of validate_opts/3.

put(TypedBucket, Key, Datatype, Spec, Opts)

-spec put(TypedBucket :: bucket_and_type(),
    Key :: binary(),
    Datatype :: datatype(),
    Spec :: type_spec(),
    Opts :: put_opts()) ->
       ok |
       {ok, babel:datatype()} |
       {ok, binary(), babel:datatype()} |
       {true | false, reliable:wf_result()} |
       {error, Reason :: any()} |
       no_return().

Transforms the datatype Datatype to a Riak Datatype using type specification Spec and stores it under TypedBucket and Key using options Opts. When called outside a workflow it returns `ok or {error, Reason}.
?> This function is workflow aware. See workflow/2 to understand the result value.

rebuild_index(Index, Collection, Opts)

-spec rebuild_index(Index :: babel_index:t(), Collection :: babel_index_collection:t(), Opts :: opts()) ->
                 {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

status(WorkRef)

-spec status(WorkRef :: reliable_work_ref:t() | binary()) ->
          {in_progress, Status :: reliable_work:status()} |
          {failed, Status :: reliable_work:status()} |
          {error, not_found | any()}.

Calls status/2.

status(WorkRef, Timeout)

-spec status(WorkRef :: reliable_work_ref:t() | binary(), Timeout :: timeout()) ->
          {in_progress, Status :: reliable_work:status()} |
          {failed, Status :: reliable_work:status()} |
          {error, not_found | any()}.

Returns the status of a Reliable Work scheduled for execution.
!> **Important** notice that at the moment completed tasks are deleted, so the abscense of a task is considered as either successful or failed, this will change in the near future as we will be retaining tasks that are discarded or completed.

type(Term)

-spec type(term()) -> set | map | counter | flag | register.

Returns the atom name for a babel datatype.

update_all_indices(Actions, Collection, Opts)

-spec update_all_indices(Actions :: [babel_index:update_action()],
                   Collection :: babel_index_collection:t(),
                   Opts :: babel_index:update_opts()) ->
                      {true | false, reliable:wf_result()} |
                      {error, Reason :: any()} |
                      no_return().

Updates all the indices in the collection that are affected by he provided Actions and schedules the update of the relevant index partitions in the database i.e. persisting the index changes.

An index in collection Collection will always be affectd in case the action is either {insert, Data} or {delete, Data} or when the action is {udpate, Old, New} and the option force was set to true or when New is not a babel map.

In case option object New is a babel map, and the option force is missing or set to false, an index will be affected by an update action only if the index's distinguished key paths have been updated or removed in the object New (See babel_index:distinguished_key_paths/1)
?> This function uses a workflow, see workflow/2 for an explanation of the possible return values.

update_indices(ActionsByIdxName, Collection, Opts)

-spec update_indices(ActionsByIdxName :: {[babel_index:update_action()], binary()},
               Collection :: babel_index_collection:t(),
               Opts :: babel_index:update_opts()) ->
                  {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Updates all the indices in the collection with the provided Actions and schedules the update of the relevant index partitions in the database i.e. persisting the index changes.

The names of the updated indices is returned under the key result of the reliable:wf_result().
?> This function uses a workflow, see workflow/2 for an explanation of the possible return values.

update_indices(Actions, IdxNames, Collection, Opts)

-spec update_indices(Actions :: [babel_index:update_action()],
               IdxNames :: [binary()],
               Collection :: babel_index_collection:t(),
               Opts :: babel_index:update_opts()) ->
                  {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Updates all the indices in the collection with the provided Actions and schedules the update of the relevant index partitions in the database i.e. persisting the index changes.

The names of the updated indices is returned under the key result of the reliable:wf_result().
?> This function uses a workflow, see workflow/2 for an explanation of the possible return values.

validate_opts(Op, Opts)

-spec validate_opts(Op :: get | put | delete, Opts :: map()) -> opts().

Equivalent to validate_opts(Op, Opts, strict).
Validates the options Opts for an operation Op.
See also: validate_opts/3.

validate_opts(Operation, Opts, Mode)

-spec validate_opts(Operation :: get | put | delete, Opts :: map(), Mode :: strict | relaxed) -> opts().

Validates the options

workflow(Fun)

-spec workflow(Fun :: fun(() -> any())) ->
            {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Equivalent to calling workflow/2 with and empty map passed as the Opts argument.
> Notice subscriptions are not working at the moment > See yield/2 to track progress.

workflow(Fun, Opts)

-spec workflow(Fun :: fun(() -> any()), Opts :: babel_workflow:opts()) ->
            {true | false, reliable:wf_result()} | {error, Reason :: any()} | no_return().

Executes the functional object Fun as a Reliable workflow, i.e. ordering and scheduling all resulting Riak KV object writes.

The code that executes inside the workflow should call one or more functions in this module to schedule writes in Riak KV. For example, if you wanted to schedule an index creation you should use create_index/2 instead of babel_index_collection, babel_index and babel_index_partition functions directly.

Any other operation, including reading and writing from/to Riak KV directly or by using the API provided by other Babel modules will work as normal and will not affect the workflow, only the special functions in this module will add work items to the workflow.

If something goes wrong inside the workflow as a result of a user error or general exception, the entire workflow is terminated and the function raises an exception. In case of an internal error, the function returns the tuple {error, Reason}.

If everything goes well, the function returns the tuple {Flag, Result}} where Flag is a boolean denoting whether a workflow was scheduled or not, and Result is a reliable:wf_result() structure.

> Notice that calling this function schedules the work to Reliable, you need to use the WorkId to check with Reliable the status of the workflow execution.

Example: Creating various babel objects and scheduling

  babel:workflow(fun() ->
   CollectionX0 = babel_index_collection:new(<<"foo">>, <<"bar">>),
   CollectionY0 = babel_index_collection:fetch(Conn, <<"foo">>, <<"users">>),
   IndexA = babel_index:new(ConfigA),
   IndexB = babel_index:new(ConfigB),
   _CollectionX1 = babel:create_index(IndexA, CollectionX0),
   _CollectionY1 = babel:create_index(IndexB, CollectionY0),
   ok
  end).
  

The resulting workflow execution will schedule the writes and deletes in the order defined by the dependency graph constructed using the results of this module functions. This ensures partitions are created first and then collections.

The Opts argument offers the following options:

* on_terminate` – a functional object `fun((Reason :: any()) -> ok). This function will be evaluated before the call terminates. In case of succesful termination the value normal will be passed as argument. Otherwise, in case of error, the error reason will be passed as argument. This allows you to perform a cleanup after the workflow execution e.g. returning a riak connection object to a pool. Notice that this function might be called multiple times in the case of nested workflows. If you need to conditionally perform a cleanup operation you might use the function is_nested_worflow/0 to take a decision.

!> **Important** notice subscriptions are not working at the moment

?> **Tip** See yield/2 to track progress.
?> **Note on Nested workflows**. No final scheduling will be done until the top level workflow is terminated. So, although a nested worflow returns {true, Result}, if the enclosing parent workflow is aborted, the entire nested workflow is aborted.

yield(WorkRef)

-spec yield(WorkRef :: reliable_worker:work_ref()) -> {ok, Payload :: any()} | timeout.

Returns the value associated with the key event_payload when used as option from a previous enqueue/2. The calling process is suspended until the work is completed or
!> **Important** notice the current implementation is not ideal as it recursively reads the status from the database. So do not abuse it. Also at the moment completed tasks are deleted, so the abscense of a task is considered as either successful or failed, this will also change as we will be retaining tasks that are discarded or completed. This will be replaced by a pubsub version soon.

yield(WorkRef, Timeout)

-spec yield(WorkRef :: reliable_worker:work_ref(), Timeout :: timeout()) ->
         {ok, Payload :: any()} | timeout.

Returns the value associated with the key event_payload when used as option from a previous enqueue/2 or timeout when Timeout milliseconds has elapsed.
!> **Important** notice The current implementation is not ideal as it recursively reads the status from the database. So do not abuse it. Also at the moment complete tasks are deleted, so the abscense of a task is considered as either succesful or failed, this will also change as we will be retaining tasks that are discarded or completed. This will be replaced by a pubsub version soon.