babel_index behaviour (babel v0.3.11)

An object that specifies the type and configuration of an application maintained index in Riak KV and the location ({bucket_type(), bucket()}, key()}) of its partitions babel_index_partition in Riak KV.

Every Index has one or more partition objects which are modelled as Riak KV maps.
An Index is persisted as a read-only CRDT Map as part of an Index Collection babel_index_collection. An Index Collection aggregates all indices for a domain entity or resource e.g. accounts.

Types

fold_fun/0

-type fold_fun() :: fun((index_key(), index_values(), any()) -> any()).

fold_opts/0

-type fold_opts() :: #{first => binary(), sort_ordering => asc | desc}.

foreach_fun/0

-type foreach_fun() :: fun((index_key(), index_values()) -> any()).

index_key/0

-type index_key() :: binary().

index_values/0

-type index_values() :: map().

key_value/0

-type key_value() :: babel_key_value:t().

local_key/0

-type local_key() :: binary().

maybe_no_return/1

-type maybe_no_return(T) :: T | no_return().

partition_id/0

-type partition_id() :: binary().

partition_key/0

-type partition_key() :: binary().

query_opts/0

-type query_opts() ::
    #{max_results => non_neg_integer() | all,
      continuation => any(),
      return_body => any(),
      timeout => timeout(),
      pagination_sort => boolean(),
      stream => boolean()}.

riak_object/0

-type riak_object() :: riakc_map:crdt_map().

t/0

-type t() ::
    #{bucket := binary(),
      bucket_type := binary(),
      config := _,
      name := binary(),
      type := atom(),
      request_opts => map()}.

update_action/0

-type update_action() ::
    {insert | delete, key_value()} | {update, Old :: key_value() | undefined, New :: key_value()}.

update_opts/0

-type update_opts() :: #{connection => pid() | fun(() -> pid()), force => boolean}.

Callbacks

distinguished_key_paths/1

-callback distinguished_key_paths(Config :: map()) -> [babel_key_value:path()].

from_riak_dict/1

-callback from_riak_dict(Dict :: orddict:orddict()) -> Config :: map().

init/2

-callback init(Name :: binary(), ConfigData :: map()) -> {ok, Config :: map()} | {error, any()}.

init_partition/2

-callback init_partition(PartitionId :: binary(), ConfigData :: map()) ->
                  {ok, babel_index_partition:t()} | {error, any()}.

iterator/3

-callback iterator(Index :: t(), Config :: map(), Opts :: babel:get_opts()) -> Iterator :: any().

iterator_done/1

-callback iterator_done(Iterator :: any()) -> boolean().

iterator_key/1

-callback iterator_key(Iterator :: any()) -> Key :: index_key().

iterator_move/3

-callback iterator_move(Action :: iterator_action(), Iterator :: any(), map()) -> any().

iterator_values/1

-callback iterator_values(Iterator :: any()) -> Key :: index_values().

match/3

-callback match(Pattern :: key_value(), babel_index_partition:t(), map()) -> [map()] | no_return().

number_of_partitions/1

-callback number_of_partitions(map()) -> pos_integer() | undefined.

partition_identifier/2

-callback partition_identifier(key_value(), map()) -> partition_id().

partition_identifiers/2

-callback partition_identifiers(asc | desc, map()) -> [partition_id()].

to_riak_object/1

-callback to_riak_object(Config :: map()) -> Object :: riak_object().

update_partition/3

-callback update_partition(Actions :: [update_action()],
                 Partition :: babel_index_partition:t(),
                 Config :: map()) ->
                    UpdatedPartition :: babel_index_partition:t().

Functions

bucket(_)

-spec bucket(t()) -> maybe_no_return(binary()).

Returns the Riak KV bucket were this index partitions are stored.

bucket_type(_)

-spec bucket_type(t()) -> maybe_no_return(binary()).

Returns the Riak KV bucket type associated with this index.

config(_)

-spec config(t()) -> maybe_no_return(riakc_map:crdt_map()).

Returns the configuration associated with this index. The configuration depends on the index type babel:type/1.

create_partitions(_)

-spec create_partitions(t()) -> [babel_index_partition:t()] | no_return().

distinguished_key_paths(Index)

-spec distinguished_key_paths(Index :: t()) -> [babel_key_value:path()].

Returns the list of the key paths that need to be present in the key value object passed as an action to update/3.

foreach(Fun, Index)

-spec foreach(foreach_fun(), Index :: t()) -> any().

from_riak_object(Object)

-spec from_riak_object(Object :: riak_object()) -> Index :: t().

match(Index, Pattern, Opts)

-spec match(Index :: t(), Pattern :: babel_index:key_value(), Opts :: babel:opts()) ->
         [{index_key(), index_values()}] | no_return().

Returns a list of matching index entries

name(_)

-spec name(t()) -> binary().

Returns name of this index

new(IndexData)

-spec new(IndexData :: map()) -> Index :: t() | no_return().

Returns a new index based on the specification map. It fails in case the specification in invalid.

A specification is map with the following fields (required fields are in bold):
**name** :: binary() – a unique name for this index within a collection. **bucket_type** :: binary() | atom() – the bucket type used to store the babel_index_partition:t() objects. This bucket type should have a datatype of map`. **bucket** :: binary() | atom() – the bucket name used to store the babel_index_partition:t() objects of this index. Typically the name of an entity in plural form e.g.accounts'. **type** :: atom() – the index type (Erlang module) used by this index. config :: map() – the configuration data for the index type used by this index.

partition_identifier(KeyValue, Index)

-spec partition_identifier(KeyValue :: key_value(), Index :: t()) -> binary().

Returns the identifier for the index partition assigned for key value object KeyValue when passed as an action to update/3.

partition_identifiers(Index)

-spec partition_identifiers(t()) -> maybe_no_return([binary()]).

Returns the list of Riak KV keys under which the partitions are stored, in ascending order. This is equivalent to the call partition_identifiers(Index, asc).

partition_identifiers(Index, Order)

-spec partition_identifiers(t(), asc | desc) -> maybe_no_return([binary()]).

Returns the list of Riak KV keys under which the partitions are stored in a defined order i.e. asc or desc.

to_delete_task(Index, PartitionId)

-spec to_delete_task(Index :: babel_index:t(), PartitionId :: binary()) -> reliable:action().

Returns the representation of this object as a Reliable Delete work item.

to_riak_object(Index)

-spec to_riak_object(Index :: t()) -> IndexCRDT :: riak_object().

to_update_task(Index, Partition)

-spec to_update_task(Index :: babel_index:t(), Partition :: babel_index_partition:t()) ->
                  reliable:action().

Returns the representation of this object as a Reliable Update work item.

type(_)

-spec type(t()) -> maybe_no_return(module()).

Returns the type of this index. A type is a module name implementing the babel_index behaviour i.e. a type of index.

typed_bucket(_)

-spec typed_bucket(t()) -> maybe_no_return({binary(), binary()}).

Returns the Riak KV typed_bucket() associated with this index.

update(Actions, Index, Opts)

-spec update(Actions :: [update_action()], Index :: t(), Opts :: update_opts()) ->
          maybe_no_return([babel_index_partition:t()]).

Throws {badaction, update_action()} in case of the action wants to delete a modified map.