babel_counter (babel v0.3.11)

Counters are a bucket-level data type that can be used by themselves, associated with a bucket/key pair, or used within a babel_map. A counter’s value can only be a positive integer, negative integer, or zero.

Types

babel_context/0

-type babel_context() :: riakc_datatype:context() | undefined | inherited.

maybe_no_return/1

-type maybe_no_return(T) :: T | no_return().

t/0
(opaque)

-opaque t()

type_spec/0

-type type_spec() :: integer.

Functions

context(T)

-spec context(T :: t()) -> babel_context().

Returns the Riak KV context

decrement(Babel_counter)

-spec decrement(t()) -> t().

Decrements the counter by 1.

decrement(Amount, Babel_counter)

-spec decrement(Amount :: integer(), t()) -> t().

Decrements the counter by amount Amount.

from_riak_counter(RiakCounter, Type)

-spec from_riak_counter(RiakCounter :: riakc_counter:counter() | integer, Type :: integer) ->
                     maybe_no_return(t()).

increment(Babel_counter)

-spec increment(t()) -> t().

Increments the counter by 1.

increment(Amount, Babel_counter)

-spec increment(Amount :: integer(), t()) -> t().

Increments the counter by amount Amount.

is_type(Term)

-spec is_type(Term :: any()) -> boolean().

is_valid_type_spec(_)

-spec is_valid_type_spec(term()) -> boolean().

new()

-spec new() -> t().

new(Value)

-spec new(Value :: integer()) -> t().

original_value(T)

-spec original_value(T :: t()) -> integer().

Returns the original value of the counter.

set(Amount, Babel_counter)

-spec set(integer(), t()) -> t().

Increments or decrements the counter so that the result of calling value/1 would be equal to amount Amount.

set_context(Ctxt, T)

-spec set_context(Ctxt :: babel_context(), T :: t()) -> NewT :: t().

This has call has no effect and it is provided for compliance withe the datatype interface.

to_riak_op(Babel_counter, _)

-spec to_riak_op(t(), type_spec()) -> riakc_datatype:update(riak_counter:coutner_op()).

type()

-spec type() -> counter.

Returns the symbolic name of this container.

value(T)

-spec value(T :: t()) -> integer().

Returns the current value of the counter.