babel_crdt (babel v0.3.11)

Types

maybe_no_return/1

-type maybe_no_return(T) :: T | no_return().

Functions

dirty_fetch(Key, Map)

-spec dirty_fetch(riakc_map:key(), riakc_map:crdt_map()) -> term().

Returns the "unwrapped" value associated with the key in the map. As opposed to riakc_map:fetch/2 this function searches for the key in the removed and updated private structures of the map first. If the key was found on the removed set, fails with a removed exception. If they key was in the updated set, it returns the updated value otherwise calls riakc_map:fetch/2.

dirty_fetch_keys(Map)

-spec dirty_fetch_keys(riakc_map:crdt_map()) -> [riakc_map:key()].

map_entry(Type, Field, Value)

-spec map_entry(Type :: riakc_datatype:typename(), Field :: binary(), Value :: binary() | list()) ->
             riakc_map:raw_entry().

register_to_atom(Unwrapped, Encoding)

-spec register_to_atom(riakc_register:register() | binary(), Encoding :: latin1 | unicode | utf8) ->
                    atom() | no_return().

register_to_binary(Unwrapped)

-spec register_to_binary(riakc_register:register() | binary()) -> binary() | no_return().

register_to_existing_atom(Unwrapped, Encoding)

-spec register_to_existing_atom(riakc_register:register() | binary(),
                          Encoding :: latin1 | unicode | utf8) ->
                             atom() | no_return().

register_to_integer(Unwrapped)

-spec register_to_integer(riakc_register:register() | binary()) -> integer() | no_return().

register_to_integer(Unwrapped, Base)

-spec register_to_integer(riakc_register:register() | binary(), Base :: 2..36) ->
                       integer() | no_return().

register_to_term(Unwrapped)

-spec register_to_term(riakc_register:register() | binary()) -> term() | no_return().

to_integer(Unwrapped)

-spec to_integer(binary() | riakc_register:register() | riakc_counter:counter()) ->
              maybe_no_return(integer()).

type(Object)

-spec type(riakc_datatype:datatype()) -> atom().