babel_flag (babel v0.3.11)

Flags behave much like Boolean values, except that instead of true or false flags have the values enable or disable.
Flags cannot be used on their own, i.e. a flag cannot be stored in a bucket/ key by itself. Instead, flags can only be stored within maps. To disable an existing flag, you have to read it or provide a context.

Types

babel_context/0

-type babel_context() :: riakc_datatype:context() | undefined | inherited.

maybe_no_return/1

-type maybe_no_return(T) :: T | no_return().

t/0
(opaque)

-opaque t()

type_spec/0

-type type_spec() :: boolean.

Functions

context(T)

-spec context(T :: t()) -> babel_context().

Returns the Riak KV context

disable(Babel_flag)

-spec disable(t()) -> t().

Disables the flag, setting its value to false.

enable(Babel_flag)

-spec enable(t()) -> t().

Enables the flag, setting its value to true.

from_riak_flag(RiakFlag, Ctxt, Type)

-spec from_riak_flag(RiakFlag :: riakc_flag:riakc_t() | boolean,
               Ctxt :: babel_context(),
               Type :: type_spec()) ->
                  maybe_no_return(t()).

is_type(Term)

-spec is_type(Term :: any()) -> boolean().

is_valid_type_spec(_)

-spec is_valid_type_spec(term()) -> boolean().

new()

-spec new() -> t().

new(Value)

-spec new(Value :: boolean()) -> t().

new(Value, Ctxt)

-spec new(Value :: boolean(), Ctxt :: babel_context()) -> t().

original_value(T)

-spec original_value(T :: t()) -> boolean().

Returns the original value of the flag.

set(_, T)

-spec set(boolean(), t()) -> t().

set_context(Ctxt, T)

-spec set_context(Ctxt :: babel_context(), T :: t()) -> NewT :: t().

Sets the context Ctxt.

to_riak_op(Babel_flag, _)

-spec to_riak_op(t(), type_spec()) -> riakc_datatype:update(riak_flag:flag_op()).

type()

-spec type() -> flag.

Returns the symbolic name of this container.

value(T)

-spec value(T :: t()) -> boolean().

Returns the current value of the flag.