babel_hash_partitioned_index (babel v0.3.11)

A hash-partitioned index is an index whose contents (index entries) have been partitioned amongst a fixed number of parts (called partitions) using a hashing algorithm to determine in which partition and entry should be located.

By partitioning an index into multiple physical parts, you are accessing much smaller objects which makes it faster and more reliable.

With hash partitioning, an index entry is placed into a partition based on the result of passing the partitioning key into a hashing algorithm.

This object is immutable.

Properties:

case_sensitive - controls wether the index terms are case sensitive or not. If true all index terms are constructed by performing a string:lowercase/1 on the values.

sort_ordering - (NOT BEING USED YET)

number_of_partitions - the number of partitions the index has. Each partition is an instance of babel_index_partition and store in Riak KV as a CRDT (the location is defined by the babel_index module properties).

partition_algorithm - the hashing algorithm used for calculating the parition for a given partition_by definition. At th moment this property only accepts jch i.e. jump consistent hashing.

partition_by - the list of keys or key paths on the target object to extract the values used for partitioning.

index_by - the list of keys or key paths on the target object to extract the values used for partitioning.

Types

fields/0

-type fields() :: [babel_key_value:key()].

iterator/0

-type iterator() :: #babel_hash_partitioned_index_iter{}.

t/0

-type t() ::
    #{case_sensitive := boolean(),
      sort_ordering := asc | desc,
      number_of_partitions := integer(),
      partition_algorithm := atom(),
      partition_identifier_prefix := binary(),
      partition_identifiers := [binary()],
      partition_by := fields(),
      index_by := fields(),
      aggregate_by := fields(),
      covered_fields := fields(),
      cardinality := one | many}.

Functions

aggregate_by(_)

-spec aggregate_by(t()) -> [binary()].

case_sensitive(_)

-spec case_sensitive(t()) -> boolean().

Returns true if the index is case sensitivity.

covered_fields(_)

-spec covered_fields(t()) -> [binary()].

distinguished_key_paths(Config)

-spec distinguished_key_paths(Config :: t()) -> [babel_key_value:key()].

from_riak_dict(Dict)

-spec from_riak_dict(Dict :: orddict:orddict()) -> Config :: t().

index_by(_)

-spec index_by(t()) -> [binary()].

init(IndexId, ConfigData)

-spec init(IndexId :: binary(), ConfigData :: map()) -> {ok, t()} | {error, any()}.

init_partition(PartitionId, ConfigData)

-spec init_partition(PartitionId :: binary(), ConfigData :: t()) ->
                  {ok, babel_index_partition:t()} | {error, any()}.

init_partitions(_)

-spec init_partitions(t()) -> {ok, [babel_index_partition:t()]} | {error, any()}.

iterator(Index, Config, Opts)

-spec iterator(babel_index:t(), babel_index:config(), Opts :: babel:opts()) -> Iterator :: iterator().

iterator_done(Iterator)

-spec iterator_done(Iterator :: any()) -> boolean().

iterator_key(Iterator)

-spec iterator_key(Iterator :: any()) -> Key :: babel_index:index_key().

iterator_move(Action, Iterator, Config)

-spec iterator_move(Action :: babel_index:iterator_action(), Iterator :: iterator(), t()) -> iterator().

iterator_values(Iterator)

-spec iterator_values(Iterator :: any()) -> Key :: babel_index:index_values().

match(Pattern, Partition, Config)

number_of_partitions(_)

-spec number_of_partitions(t()) -> pos_integer().

partition_algorithm(_)

-spec partition_algorithm(t()) -> atom().

Returns the partition algorithm name configured for this index.

partition_by(_)

-spec partition_by(t()) -> [binary()].

partition_identifier(KeyValue, Config)

-spec partition_identifier(babel_index:key_value(), t()) -> babel_index:partition_id() | no_return().

partition_identifier_prefix(_)

-spec partition_identifier_prefix(t()) -> binary().

partition_identifiers(Order, Config)

-spec partition_identifiers(asc | desc, t()) -> [babel_index:partition_id()].

sort_ordering(_)

-spec sort_ordering(t()) -> asc | desc.

Returns the sort ordering configured for this index. The result can be the atoms asc or desc.

to_riak_object(Config)

-spec to_riak_object(Config :: t()) -> ConfigCRDT :: babel_index:riak_object().

update_partition(ActionData, Partition, Config)

-spec update_partition(ActionData :: babel_index:update_action() | [babel_index:update_action()],
                 Partition :: babel_index_partition:t(),
                 Config :: t()) ->
                    babel_index_partition:t() | no_return().