babel_index_collection (babel v0.3.11)

A babel_index_collection is a Riak Map representing a mapping from binary keys to babel_index objects, where keys are the value of the babel_index:name/1 property.

An Index Collection has a name that typically represents the name of a resource (or entity) name in your domain model e.g. accounts, users.

A babel collection object is stored in Riak KV under a bucket_type that should be defined through configuration using the index_collection_bucket_type configuration option; and a bucket name which results from concatenating a prefix provided as argument in this module functions and the suffix "/index_collection".

## Configuring the bucket type

The bucket type needs to be configured and activated in Riak KV before using this module. The datatype property of the bucket type should be configured to map.

The following example shows how to configure and activate the bucket type with the recommeded default replication properties, for the example we asume the application property index_collection_bucket_type maps to "index_collection" bucket type name.

     shell
  riak-admin bucket-type create index_collection '{"props":
  {"datatype":"map",
  "n_val":3, "pw":"quorum", "pr":"quorum", "notfound_ok":false,
  "basic_quorum":true}}'
  riak-admin bucket-type activate index_collection

## Default replication properties
All functions in this module resulting in reading or writing to Riak KV allow an optional map with Riak KV's replication properties, but we recommend to use of the functions which provide the default replication properties.

Types

fold_fun/0

-type fold_fun() :: fun((Key :: key(), Value :: any(), AccIn :: any()) -> AccOut :: any()).

key/0

-type key() :: binary().

riak_object/0

-type riak_object() :: riakc_map:crdt_map().

t/0

-type t() :: #babel_index_collection{}.

Functions

add_index(Index, Collection)

-spec add_index(Index :: babel_index:t(), Collection :: t()) -> t() | no_return().

Returns a copy of collection Collection where the index Index has been added. If the an index with the same name existed in the collection, it will be replaced by Index.
!> **Important**. This is a private API. If you want to add an index to the collection and create the index in Riak KV use babel:create_index/3 instead.

bucket(Collection)

-spec bucket(Collection :: t()) -> binary().

data(Collection)

-spec data(Collection :: t()) -> orddict:orddict().

delete(BucketPrefix, Key, Opts)

-spec delete(BucketPrefix :: binary(), Key :: binary(), Opts :: babel:delete_opts()) ->
          ok | {error, not_found | term()}.

delete_index(Id, Collection)

-spec delete_index(Id :: binary() | babel_index:t(), Collection :: t()) -> t() | no_return().

fetch(BucketPrefix, Key, Opts)

-spec fetch(BucketPrefix :: binary(), Key :: binary(), Opts :: babel:get_opts()) -> t() | no_return().

fold(Fun, Acc, Collection)

-spec fold(Fun :: fold_fun(), Acc :: any(), Collection :: t()) -> any().

from_riak_object(Object)

-spec from_riak_object(Object :: riak_object()) -> Collection :: t().

id(Collection)

-spec id(Collection :: t()) -> binary().

index(IndexName, Collection)

-spec index(IndexName :: binary(), Collection :: t()) -> babel_index:t() | no_return().

Returns the babel index associated with name IndexName in collection Collection. This function assumes that the name is present in the collection. An exception is generated if it is not.

index_names(Collection)

-spec index_names(Collection :: t()) -> [binary()].

indices(Collection)

-spec indices(Collection :: t()) -> [babel_index:t()].

Returns all the indices in the collection.

is_index(IndexName, Collection)

-spec is_index(IndexName :: binary(), Collection :: t()) -> babel_index:t() | no_return().

lookup(BucketPrefix, Key, Opts)

-spec lookup(BucketPrefix :: binary(), Key :: binary(), Opts :: babel:get_opts()) ->
          {ok, t()} | {error, not_found | term()}.

new(BucketPrefix, Name)

-spec new(BucketPrefix :: binary(), Name :: binary()) -> t().

Creates a new index collection object. The value for bucket is computed by concatenating BucketPrefix with the suffix /index_collection.

size(Collection)

-spec size(Collection :: t()) -> non_neg_integer().

Returns the number of elements in the collection Collection.

store(Collection, Opts)

-spec store(Collection :: t(), Opts :: babel:put_opts()) ->
         {ok, Index :: t()} | {error, Reason :: any()}.

Stores an index collection in Riak KV. The collection will be stored under the bucket type configured for the application option index_collection_bucket_type, bucket name will be the value returned by bucket/1, and the key will be the value returned by id/1.

to_delete_task(Collection)

-spec to_delete_task(Collection :: t()) -> reliable:action().

to_riak_object(Collection)

-spec to_riak_object(Collection :: t()) -> Object :: riak_object().

to_update_task(Collection)

-spec to_update_task(Collection :: t()) -> reliable:action().