babel_key_value (babel v0.3.11)

A Key Value coding interface for property lists and maps.

Types

fold_fun/0

-type fold_fun() :: fun((Key :: key(), Value :: any(), AccIn :: any()) -> AccOut :: any()).

key/0

-type key() :: atom() | binary() | tuple() | riakc_map:key() | path().

path/0

-type path() :: [atom() | binary() | tuple() | riakc_map:key()].

t/0

-type t() :: map() | [proplists:property()] | babel_map:t() | riakc_map:crdt_map().

Functions

collect(Keys, KVTerm)

-spec collect([key()], KVTerm :: t()) -> [any()].

collect(Keys, KVTerm, Default)

-spec collect([key()], KVTerm :: t(), Default :: any()) -> [any()].

fold(Fun, Acc, KVTerm)

-spec fold(Fun :: fold_fun(), Acc :: any(), KVTerm :: t()) -> NewAcc :: any().

get(Key, KVTerm)

-spec get(Key :: key(), KVTerm :: t()) -> Value :: term().

Returns value Value associated with Key if KVTerm contains Key. Key can be an atom, a binary or a path represented as a list of atoms and/ or binaries.

The call fails with a {badarg, KVTerm} exception if KVTerm is not a property list, map or Riak CRDT Map. It also fails with a {badkey, Key} exception if no value is associated with Key.
> In the case of Riak CRDT Maps a key MUST be a riakc_map:key().

get(Key, KVTerm, Default)

-spec get(Key :: key(), KVTerm :: t(), Default :: term()) -> term().

put(Key, Value, KVTerm)

-spec put(Key :: key(), Value :: any(), KVTerm :: t()) -> t().

Equivalent to set / 3.

set(Key, Value, KVTerm)

-spec set(Key :: key(), Value :: any(), KVTerm :: t()) -> t().