babel_map (babel v0.3.11)

Provides an alternative to Riak Map Datatype.

# Overview

Babel maps (maps) differ from Riak's and Erlang's maps in several ways:

* Maps are special key-value structures where the key is a binary name and the value is a Babel datatype, each one an alternative of the Riak's counterparts, with the exception of the Riak Register type which can be represented by any Erlang Term in Babel (and not just a binary) provided there exists a valid type conversion specification (see Type Specifications). * Maps maintain the same semantics as Riak Maps but with some key differences * As with Riak Map, removals, and modifications are captured for later application by Riak but they are also applied to the local state. That is, maps resolve the issue of not being able to read your object mutations in memory that occurs when using Riak maps. * Removals are processed before updates in Riak. Also, removals performed without a context may result in failure. * Updating an entry followed by removing that same entry will result in no operation being recorded. Likewise, removing an entry followed by updating that entry will cancel the removal operation. * You may store or remove values in a map by using set/3`, `remove/2, and other functions targetting embedded babel containers e.g. add_element/ 3, add_elements/3, del_element/3 to modify an embeded babel_set . This is a complete departure from Riak's cumbersome update/3 function. As in Riak Maps, setting or adding a value to an embedded container that is not present will create a new container before the set/add operation. * Certain function e.g. set/3 allows you to set a value in a key or a path (list of nested keys).

# Type Specifications

A type specification is an Erlang map where the keys are the Babel map keys and their value is another specification or a type_mapping().

For example the specification #{<<"friends">> => {set, list}}, says the map contains a single key name "friends" containing a Babel Set (compatible with Riak Set) where the individual elements are represented in Erlang as lists i.e. a set of lists. This will result in a map containing the key <<"friends">> and a babel set contining the elements converted from binaries to lists.
The special '\_' key name provides the capability to convert a Riak Map where the keys are not known in advance. These specs can only have a single entry as follows #{{\_', set}, erl_type()}'.

Types

action/0

-type action() :: map().

babel_context/0

-type babel_context() :: riakc_datatype:context() | undefined | inherited.

collect_opts/0

-type collect_opts() :: #{default => any(), on_badkey => skip | error, return => map | list}.

datatype/0

-type datatype() :: counter | flag | register | set | map.

erl_type/0

-type erl_type() ::
    atom | existing_atom | boolean | integer | float | binary | list |
    fun((encode, any()) -> value()) |
    fun((decode, value()) -> any()).

key/0

-type key() :: binary().

key_path/0

-type key_path() :: binary() | [binary()].

maybe_no_return/1

-type maybe_no_return(T) :: T | no_return().

t/0
(opaque)

-opaque t()

type_mapping/0

-type type_mapping() ::
    {map, type_spec()} |
    {set, erl_type()} |
    {counter, erl_type()} |
    {flag, erl_type()} |
    {register, erl_type()}.

type_spec/0

-type type_spec() :: #{'$validated' => true, key() | '_' => type_mapping()}.

type_spec_ref/0

-type type_spec_ref() :: {type_spec_ref, any()}.

value/0

-type value() :: any().

Functions

add_element(Key, Value, Map)

-spec add_element(Key :: key_path(), Value :: value(), Map :: t()) -> NewMap :: maybe_no_return(t()).

Adds element Value to a babel set associated with key or path Key in map Map and inserts the association into map NewMap.

If the key Key does not exist in map Map, this function creates a new babel set containining Value.

The call might fail with the following exception reasons:
* {badset, Set} - if the initial value associated with Key in map Map0 is not a babel set; * {badmap, Map} exception if Map is not a babel map. * {badkey, Key} - exception if no value is associated with Key or Key is not of type binary.

add_elements(Key, Values, Map)

-spec add_elements(Key :: key_path(), Values :: [value()], Map :: t()) -> NewMap :: maybe_no_return(t()).

Adds a list of values Values to a babel set associated with key or path Key in map Map and inserts the association into map NewMap.

If the key Key does not exist in map Map, this function creates a new babel set containining Values.

The call might fail with the following exception reasons:
* {badset, Set} - if the initial value associated with Key in map Map0 is not a babel set; * {badmap, Map} exception if Map is not a babel map. * {badkey, Key} - exception if no value is associated with Key or Key is not of type binary.

change_status(KeyOrPath, Map)

-spec change_status(KeyOrPath :: key_path(), Map :: t()) -> none | both | removed | updated | any().

Returns the status of a key path KeyPath in map Map, where status can be one of updated, removed, both or none.

change_status(KeyOrPath, Map, Default)

-spec change_status(KeyOrPath :: key_path(), Map :: t(), Default :: any()) ->
                 none | removed | updated | any() | no_return().

Returns the status of a key path KeyPath in map Map, where status can be one of updated, removed, both or none.

changed_key_paths(T)

-spec changed_key_paths(T :: t()) -> {Updated :: [key_path()], Removed :: [key_path()]} | no_return().

Returns a tuple where the first element is the list of the key paths that have been updated and the second one those which have been removed in map T. Notice that a key path might be both removed and updated, in which case it will be a mamber of both result elements. The call fails with a {badmap, T} exception if T is not a map.

collect(Keys, Map)

-spec collect([key_path()], Map :: t()) -> [any()].

Calls collect/3 with the default options.

collect(Keys, Map, Opts)

-spec collect(Keys :: [key_path()], Map :: t(), Opts :: collect_opts()) ->
           [any()] | #{binary() => any()}.

Returns a list of values associated with the keys Keys.

?> The value returned by this function are not raw values, but Babel datatype values. If you want to get the raw values use collect_values/3 instead.

The return depends on the following options:

* default - the value to use as default when a key in Keys is not present in the map Map. The presence of a default value disables the option on_badkey. * on_badkey - what happens when a key is not present in the map and there was no default value provided. Valid values are skip, or error. When using skip the function simply ignores the missing key and returns all found keys. Using error will fail with a badkey exception. * return` - the Erlang return type of the function. Valid values are `list and map. Notice that naturally Erlang maps will deduplicate keys whereas lists would not. Default value: list.

**Examples**:

  Map = babel_map:new(
      #{
          <<"x">> => #{
              <<"a">> => 1,
              <<"b">> => 2
          }
      },
      #{
          <<"x">> => {map, #{
              <<"a">> => {counter, integer},
              <<"b">> => {counter, integer}
          }}
      }
  ).
  

  babel_map:collect([<<"x">>], Map).
  [{babel_map,#{<<"a">> => {babel_counter,0,1},
                <<"b">> => {babel_counter,0,2}},
              [<<"a">>,<<"b">>],
              [],undefined}]
  

  babel_map:collect([<<"y">>], Map).
  ** exception error: badkey
  

  babel_map:collect([<<"y">>], Map, #{on_badkey => skip}).
  []
  

  babel_map:collect([<<"y">>], Map, #{default => undefined}).
  [undefined]
  

  babel_map:collect([<<"x">>], Map, #{return => map}).
  #{<<"x">> =>
        {babel_map,#{<<"a">> => {babel_counter,0,1},
                     <<"b">> => {babel_counter,0,2}},
                   [<<"a">>,<<"b">>],
                   [],undefined}}
  

  babel_map:collect(
      [ [<<"x">>, <<"a">>], [<<"x">>, <<"b">>]  ],
      Map,
      #{return => list}
  ).
  [{babel_counter, 0, 1},{babel_counter, 0, 2}]
  

  babel_map:collect(
      [ [<<"x">>, <<"a">>], [<<"x">>, <<"b">>]  ],
      Map,
      #{return => map}
  ).
  #{<<"x">> =>
        #{<<"a">> => {babel_counter,0,1},
          <<"b">> => {babel_counter,0,2}}}
  
!> The function is not clever in terms of optimisations, so judgment is required when used. For example if Keys = [ [A, B, X], [A, B, Y], [A, B, Z] ], it will iterate 3 times traversing the whole path from A to X, Y and Z i.e. reading A then B three times. In the future we might want to change this so that [A, B] is read once.

collect_values(Keys, Map)

-spec collect_values([key_path()], Map :: t()) -> [any()].

Returns a list of values associated with the keys Keys. Fails with a {badkey, K}` exeception if any key `K in Keys is not present in the map.

collect_values(Keys, Map, Opts)

-spec collect_values([key_path()], Map :: t(), Opts :: collect_opts()) -> [any()] | #{binary() => any()}.

Returns a list of values associated with the keys Keys.

?> The value returned by this function are the raw values e.g. equivalent to calling get_value/2 on a Babel datatype. If you want to get the container datatype values use collect/3 instead.

The return depends on the following options:

* default - the value to use as default when a key in Keys is not present in the map Map. The presence of a default value disables the option on_badkey. * on_badkey - what happens when a key is not present in the map and there was no default value provided. Valid values are skip, or error. When using skip the function simply ignores the missing key and returns all found keys. Using error will fail with a badkey exception. * return` - the Erlang return type of the function. Valid values are `list and map. Notice that naturally Erlang maps will deduplicate keys whereas lists would not. Default value: list.

**Examples**:

  Map = babel_map:new(
      #{
          <<"x">> => #{
              <<"a">> => 1,
              <<"b">> => 2
          }
      },
      #{
          <<"x">> => {map, #{
              <<"a">> => {counter, integer},
              <<"b">> => {counter, integer}
          }}
      }
  ).
  

  babel_map:collect_values([<<"x">>], Map).
  [#{<<"a">> => 1, <<"b">> => 2}]
  

  babel_map:collect_values([<<"y">>], Map).
  ** exception error: badkey
  

  babel_map:collect_values([<<"y">>], Map, #{on_badkey => skip}).
  []
  

  babel_map:collect_values([<<"y">>], Map, #{default => undefined}).
  [undefined]
  

  babel_map:collect_values([<<"x">>], Map, #{return => map}).
  #{<<"x">> => #{<<"a">> => 1, <<"b">> => 2}}.
  

  babel_map:collect_values(
      [ [<<"x">>, <<"a">>], [<<"x">>, <<"b">>]  ],
      Map,
      #{return => list}
  ).
  [1,2]
  

  babel_map:collect_values(
      [ [<<"x">>, <<"a">>], [<<"x">>, <<"b">>]  ],
      Map,
      #{return => map}
  ).
  #{<<"x">> => #{<<"a">> => 1, <<"b">> => 2}}
  
!> The function is not clever in terms of optimisations, so judgment is required when used. For example if Keys = [ [A, B, X], [A, B, Y], [A, B, Z] ], it will iterate 3 times traversing the whole path from A to X, Y and Z i.e. reading A then B three times. In the future we might want to change this so that [A, B] is read once.

context(T)

-spec context(T :: t()) -> babel_context() | no_return().

Returns the Riak KV context associated with map T. The call fails with a {badmap, T} exception if T is not a map.

decrement(Key, T)

-spec decrement(Key :: key_path(), T :: t()) -> NewT :: t().

decrement(Key, Value, T)

-spec decrement(Key :: key_path(), Value :: integer(), T :: t()) -> NewT :: t().

del_element(Key, Value, Map)

-spec del_element(Key :: key_path(), Value :: value(), Map :: t()) -> NewMap :: maybe_no_return(t()).

Returns a new map NewMap were the value Value has been removed from a babel set associated with key or path Key in map Map.

If the key Key does not exist in map Map, this function creates a new babel set recording the removal of Value.

The call might fail with the following exception reasons:
* {badset, Set} - if the initial value associated with Key in map Map0 is not a babel set; * {badmap, Map} exception if Map is not a babel map. * {badkey, Key} - exception if no value is associated with Key or Key is not of type binary.

disable(Key, T)

-spec disable(Key :: key_path(), T :: t()) -> NewT :: t().

enable(Key, T)

-spec enable(Key :: key_path(), T :: t()) -> NewT :: t().

find(Key, T)

-spec find(Key :: key_path(), T :: t()) -> {ok, any()} | error.

Returns the tuple {ok, Value :: any()} if the key 'Key' is associated with value Value in map T. Otherwise returns the atom error. The call fails with a {badmap, T} exception if T is not a map and {badkey, Key} exception if Key is not a binary term.

from_riak_map(RMap, Spec)

-spec from_riak_map(RMap :: riakc_map:crdt_map() | list(), Spec :: type_spec()) -> t().

Returns a new map by applying the type specification Spec to the Riak Map RMap.

from_riak_map(RMap, Spec, Opts)

-spec from_riak_map(RMap :: riakc_map:crdt_map() | list(),
              Spec :: type_spec() | type_spec_ref(),
              Opts :: map()) ->
                 t().

Returns a new map by applying the type specification Spec to the Riak Map RMap.
Options is a map containing th following options: * missing_spec -- the strategy to deal with a field found in the Riak Map that does not have a mapping in the type specificiation Spec. It accepts the values error` or `ignore`. In the case of `error` a `{missing_spec, Key} exception will be raised; in the case or ignore it will simply be ignored. The default is error.

get(Key, T)

-spec get(Key :: key_path(), T :: t()) -> any() | no_return().

Returns value Value associated with Key if T contains Key. Key can be a binary or a path represented as a list of binaries.
The call fails with a {badarg, T} exception if T is not a Babel Map. It also fails with a {badkey, Key} exception if no value is associated with Key or if Key is not a binary term.

get(Key, Map, Default)

-spec get(Key :: key_path(), Map :: t(), Default :: any()) -> Value :: value().

Returns value Value associated with Key if T contains Key, or the default value Default in case T does not contain Key.

Key can be a binary or a path represented as a list of binaries.
The call fails with a {badarg, T}` exception if `T is not a Babel Map. It also fails with a {badkey, Key}` exception if no value is associated with `Key or if Key is not a binary term.

get_registered_type_spec(Ref)

-spec get_registered_type_spec(Ref :: any()) -> type_spec_ref() | undefined.

get_value(Key, T)

-spec get_value(Key :: key_path(), T :: t()) -> any().

An util function equivalent to calling DatatypeMod:value(get(Key, T)).

get_value(Key, T, Default)

-spec get_value(Key :: key_path(), T :: t(), Default :: any()) -> any() | no_return().

An util function equivalent to calling DatatypeMod:value(get(Key, T, Default)).

has_changes(Map)

-spec has_changes(Map :: t()) -> boolean().

Returns true if at least one property of the map has been updated (or inserted) or removed and false otherwise.

increment(Key, T)

-spec increment(Key :: key_path(), T :: t()) -> NewT :: t().

increment(Key, Value, T)

-spec increment(Key :: key_path(), Value :: integer(), T :: t()) -> NewT :: t().

is_type(Term)

-spec is_type(Term :: any()) -> boolean().

Returns true if term Term is a babel map. The call fails with a {badmap, Term} exception if Term is not a map.

keys(T)

-spec keys(T :: t()) -> [binary()] | no_return().

Returns a complete list of keys, in any order, which resides within map T. The call fails with a {badmap, T} exception if T is not a map.

new()

-spec new() -> t().

Creates a new empty map.

new(Data)

-spec new(Data :: map()) -> t().

Creates a new map from the erlang map Data, previously filtering out all keys assigned to the undefined. This function converts the erlang types map(), list() and boolean() to their corresponding Babel Datatypes babel_map:t(), babel_map:set() and babel_map:flag(). Any other value will be assumed to be a register. Also, there is not type validation or coersion when creating a babel_set:t() out of a list.
!> **Important**. Notice that using this function might result in incompatible types when later using a type specification e.g. to_riak_op/2. We strongly suggest not using this function and using new/2 instead.

new(Data, Spec)

-spec new(Data :: map(), Spec :: type_spec()) -> t().

Equivalent to new(Data, Spec, undefined).

new(Data, Spec, Ctxt)

-spec new(Data :: map(), Spec :: type_spec() | type_spec_ref(), Ctxt :: riakc_datatype:context()) -> t().

Creates a new Babel Map from the erlang map Data, previously filtering out all keys assigned to the undefined.

patch(ActionList, T)

-spec patch(ActionList :: [action()], T :: t()) -> NewT :: t().

Updates a map T with the provide key-value action list ActionList. If the value associated with a key Key in Values is equal to undefined` this equivalent to calling `remove(Key, Map) with the difference that an exception will not be raised in case the map had no context assigned.
This function fails with missing_spec if it does not have a type specification reference. See update/3 to pass a type specification.

patch(ActionList, T, Spec)

-spec patch(ActionList :: [action()], T :: t(), Spec :: type_spec()) -> NewT :: t().

Updates a map T with the provide key-value action list ActionList. If the value associated with a key Key in Values is equal to undefined` this equivalent to calling `remove(Key, Map) with the difference that an exception will not be raised in case the map had no context assigned.
Example:

put(Key, Value, Map)

-spec put(Key :: key_path(), Value :: value(), Map :: t()) -> NewMap :: maybe_no_return(t()).

Same as set/3.

register_type_spec(Spec)

-spec register_type_spec(type_spec()) -> type_spec_ref().

Equivalent to register_type_spec(undefined, Spec).

Registers a type specification Spec and returns a type_spec_ref() that can be use with any other map instance of the same kind. A registered type spec is validated during registration and the registration fails if it does not pass the validation. All functions that take a type specification as argument also accept a type specification reference, which is more efficient as there is no need to have the type specification in the local process memery and there is no need for further validation.

This call fails with exception {invalid_spec, Errors :: map()} if the type specification Spec is invalid.
The storage for specifications is implemented using persistent_term. So refer to that module's documentation to understand best practices. In particular, it is ideal to register the type specifications as soon as your aplication starts in order to minimize the number of processes on the node before performing a registration. It would also be wise to avoid registering type specifications when the system is at peak load.

register_type_spec(Id, Spec0)

-spec register_type_spec(Id :: any() | undefined, type_spec()) -> type_spec_ref().

Registers a type specification Spec under a reference using unique identifier Id. Returns a type_spec_ref(). If Id equals undefined, then id is computed by computing a hash of value of Spec. The returned reference can be use with any other map instance of the same kind. A registered type spec is validated during registration and the registration fails if it does not pass the validation. All functions that take a type specification as argument also accept a type specification reference, which is more efficient as there is no need to have the type specification in the local process memery and there is no need for further validation.

This call fails with exception {invalid_spec, Errors :: map()} if the type specification Spec is invalid.
The storage for specifications is implemented using persistent_term. So refer to that module's documentation to understand best practices. In particular, it is ideal to register the type specifications as soon as your aplication starts in order to minimize the number of processes on the node before performing a registration. It would also be wise to avoid registering type specifications when the system is at peak load.

remove(Key, T)

-spec remove(Key :: key_path(), T :: t()) -> NewT :: maybe_no_return(t()).

Removes a key and its value from the map. Removing a key that does not exist simply records a remove operation.
In case the map has no context i.e. it is newly created, removes will only %% work for keys that have been updated since creation. In all other cases the call will fail with a context_required exception.

set(Key, Value, Map)

-spec set(Key :: key_path(), Value :: value(), Map :: t()) -> NewMap :: maybe_no_return(t()).

Associates Key with value Value and inserts the association into map NewMap. If key Key already exists in map Map, the old associated value is replaced by value Value. The function returns a new map NewMap containing the new association and the old associations in Map.

Passing a Value of undefined` is equivalent to calling `remove(Key, Map) with the difference that an exception will not be raised in case the map had no context assigned.
The call fails with a {badmap, Term} exception if Map or any value of a partial key path is not a babel map.

set_context(Ctxt, T)

-spec set_context(Ctxt :: riakc_datatype:context(), T :: t()) -> NewT :: t().

Sets the context Ctxt.

set_elements(Key, Values, Map)

-spec set_elements(Key :: key_path(), Values :: [value()], Map :: t()) -> NewMap :: maybe_no_return(t()).

Sets a list of values Values to a babel set associated with key or path Key in map Map and inserts the association into map NewMap. See babel_set:set_elements/2.

If the key Key does not exist in map Map, this function creates a new babel set containining Values.

The call might fail with the following exception reasons:
* {badset, Set} - if the initial value associated with Key in map Map0 is not a babel set; * {badmap, Map} exception if Map is not a babel map. * {badkey, Key} - exception if no value is associated with Key or Key is not of type binary.

size(T)

-spec size(T :: t()) -> non_neg_integer() | no_return().

Returns the size of the values of the map T. The call fails with a {badmap, T} exception if T is not a map.

to_riak_op(Babel_map)

Extracts a Riak Operation from the map to be used with a Riak Client update request. The call fails with a {badmap, T} exception if T is not a map and missing_spec if it does not have a type specification reference.

to_riak_op(T, Spec)

-spec to_riak_op(T :: t(), Spec :: type_spec()) ->
              riakc_datatype:update(riakc_map:map_op()) | no_return().

Extracts a Riak Operation from the map to be used with a Riak Client update request. The call fails with a {badmap, T} exception if T is not a map.

type()

-spec type() -> map.

Returns the symbolic name of this container.

type_spec_ref(T)

-spec type_spec_ref(T :: t()) -> type_spec_ref() | undefined.

Returns the type specification reference associated with map T or undefined is there is none.

unregister_type_spec(Ref)

-spec unregister_type_spec(Ref :: any()) -> boolean().

update(Values, T)

-spec update(Values :: babel_key_value:t(), T :: t()) -> NewT :: t().

Updates a map T with the provide key-value pairs Values. If the value associated with a key Key in Values is equal to undefined` this equivalent to calling `remove(Key, Map) with the difference that an exception will not be raised in case the map had no context assigned.
This function fails with missing_spec if it does not have a type specification reference. See update/3 to pass a type specification.

update(Values, T, Spec)

-spec update(Values :: babel_key_value:t(), T :: t(), Spec :: type_spec()) -> NewT :: t().

Updates a map T with the provide key-value pairs Values. If the value associated with a key Key in Values is equal to undefined` this equivalent to calling `remove(Key, Map) with the difference that an exception will not be raised in case the map had no context assigned.

validate_type_spec(Spec)

-spec validate_type_spec(Spec :: type_spec()) -> type_spec() | no_return().

value(Map)

-spec value(Map :: t()) -> map() | no_return().

Returns an external representation of the map Map as an Erlang map(). This is build recursively by calling the value/1 function on any embedded datatype. The call fails with a {badmap, T} exception if T is not a map.