babel_set (babel v0.3.11)

Types

babel_context/0

-type babel_context() :: riakc_datatype:context() | undefined | inherited.

maybe_no_return/1

-type maybe_no_return(T) :: T | no_return().

t/0
(opaque)

-opaque t()

type_spec/0

-type type_spec() ::
    atom | existing_atom | boolean | integer | float | binary | list |
    fun((encode, any()) -> binary()) |
    fun((decode, binary()) -> any()).

Functions

add_element(Element, T)

-spec add_element(Element :: any(), T :: t()) -> t().

Adds an element to the set. You may add an element that already exists in the original set value, but it does not count for the object's size calculation. Adding an element that already exists is non-intuitive, but acts as a safety feature: a client code path that requires an element to be present in the set (or removed) can ensure that intended state by applying an operation.

add_elements(Elements, T)

-spec add_elements(Elements :: [any()], T :: t()) -> t().

context(T)

-spec context(T :: t()) -> babel_context().

Returns the Riak KV context

del_element(Element, T)

-spec del_element(Element :: any(), T :: t()) -> t() | no_return().

Removes an element from the set. You may remove an element that does not appear in the original set value. This is non-intuitive, but acts as a safety feature: a client code path that requires an element to be present in the set (or removed) can ensure that intended state by applying an operation.

del_elements(Elements, T)

-spec del_elements(Elements :: [any()], T :: t()) -> t() | no_return().

Removes an element from the set. You may remove an element that does not appear in the original set value. This is non-intuitive, but acts as a safety feature: a client code path that requires an element to be present in the set (or removed) can ensure that intended state by applying an operation.

fold(Fun, Acc, T)

-spec fold(Fun :: fun((term(), term()) -> term()), Acc :: term(), T :: t()) -> Acc :: term().

Folds over the members of the set.

from_riak_set(RiakSet, Type)

-spec from_riak_set(RiakSet :: riakc_set:riakc_set() | ordsets:ordset(), Type :: type_spec()) ->
                 maybe_no_return(t()).

is_element(Element, Babel_set)

-spec is_element(binary(), t()) -> boolean().

Test whether an element is a member of the set.

is_original_element(Element, Babel_set)

-spec is_original_element(binary(), t()) -> boolean().

Test whether an element is a member of the original set i,e. the one retrieved from Riak.

is_type(Term)

-spec is_type(Term :: any()) -> boolean().

is_valid_type_spec(Fun)

-spec is_valid_type_spec(term()) -> boolean().

new()

-spec new() -> t().

new(Data)

-spec new(Data :: ordsets:ordset(any())) -> t().

!> **Important**. Notice that using this function might result in incompatible types when later using a type specification e.g. to_riak_op/2. We strongly suggest not using this function and using new/2 instead.

new(Data, Type)

-spec new(Data :: ordsets:ordset(any()), Type :: type_spec()) -> t().

original_value(Babel_set)

-spec original_value(t()) -> ordsets:ordset(any()).

Returns the original value of the set as an ordset. This is equivalent to riakc_set:value/1 but where the elements are binaries but of the type defined by the conversion spec() used to create the set.

set_context(Ctxt, T)

-spec set_context(Ctxt :: babel_context(), T :: t()) -> NewT :: t().

Sets the context Ctxt.

set_elements(Elements, T)

-spec set_elements(Elements :: [any()], T :: t()) -> NewT :: t().

size(T)

-spec size(T :: t()) -> pos_integer().

Returns the cardinality (size) of the set.

subtract(Set1, Set2)

-spec subtract(Set1 :: t(), Set2 :: t()) -> Set3 :: t().

Returns a copy of Set1 containing only the elements of Set1 that are not also elements of Set2.

to_riak_op(T, Type)

-spec to_riak_op(T :: t(), Type :: type_spec()) -> riakc_datatype:update(riakc_set:set_op()).

type()

-spec type() -> set.

Returns the symbolic name of this container.

type_spec(Term)

-spec type_spec(Term :: any()) -> type_spec() | undefined.

value(T)

-spec value(T :: t()) -> ordsets:ordset(any()).

Returns the current value of the set.